refactored suggesters, flags; call concordances before filtering
